seperate cursive step
