+++ src/parse/recursion.js
@@ -1,0 +1,58 @@
+const i18n = require('../data/i18n');
+const parse_infobox = require('./infobox');
+const parse_infobox_template = require('./infobox_template');
+const parse_image = require('./image');
+const languages = require('../data/languages');
+const recursive_matches = require('./cleanup/recursive_matches');
+const infobox_reg = new RegExp('{{(' + i18n.infoboxes.join('|') + ')[: \n]', 'ig');
+const fileRegex = new RegExp('\\[\\[(' + i18n.images.concat(i18n.files).join('|') + ')', 'i');
+const noWrap_reg = /^\{\{nowrap\|(.*?)\}\}$/;
+
+//reduce the scary recursive situations
+const parse_recursive = function(r, wiki) {
+  //remove {{template {{}} }} recursions
+  let matches = recursive_matches('{', '}', wiki);
+  matches.forEach(function(s) {
+    if (s.match(infobox_reg, 'ig') && Object.keys(r.infobox).length === 0) {
+      r.infobox = parse_infobox(s);
+      r.infobox_template = parse_infobox_template(s);
+    }
+    if (s.match(infobox_reg)) {
+      wiki = wiki.replace(s, '');
+    }
+    //rest of them...
+    if (s.match(/^\{\{/)) {
+      //support nowrap
+      const nowrap = s.match(noWrap_reg);
+      if (nowrap) {
+        wiki = wiki.replace(s, nowrap[1]);
+        return;
+      }
+      //if it's not a known template, but it's recursive, remove it
+      //(because it will be misread later-on)
+      wiki = wiki.replace(s, '');
+    }
+  });
+
+  //second, remove [[file:...[[]] ]] recursions
+  matches = recursive_matches('[', ']', wiki);
+  matches.forEach(function(s) {
+    if (s.match(fileRegex)) {
+      r.images.push(parse_image(s));
+      wiki = wiki.replace(s, '');
+    }
+  });
+  //third, wiktionary-style interlanguage links
+  matches.forEach(function(s) {
+    if (s.match(/\[\[([a-z][a-z]):(.*?)\]\]/i) !== null) {
+      const lang = s.match(/\[\[([a-z][a-z]):/i)[1];
+      if (lang && languages[lang]) {
+        r.translations[lang] = s.match(/\[\[([a-z][a-z]):(.*?)\]\]/i)[2];
+      }
+      wiki = wiki.replace(s, '');
+    }
+  });
+  return wiki;
+};
+
+module.exports = parse_recursive;
